Fixed numerous GNU/Linux compile issues
git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/opal/trunk@27656 023b2edf-31b2-4de3-b41e-bca80c47788f

+++ plugins/video/MPEG4-ffmpeg/mpeg4.cxx
@@ -61,12 +61,13 @@
 
 #include <codec/opalplugin.hpp>
 
-#include "../../common/ffmpeg.h"
+#include "../common/ffmpeg.h"
 
 
 // Needed C++ headers
 #include <deque>
 #include <vector>
+
 
 using namespace std;
 
@@ -459,7 +460,7 @@ class MPEG4_EncodedFrame : public FFMPEGCodec::EncodedFrame
 {
   protected:
     // packet sizes generating in RtpCallback
-    deque<unsigned> m_packetSizes;
+    deque<size_t> m_packetSizes;
     unsigned m_fragmentationOffset;
 
   public:
